primary Color and secondary color set of as type UIColor
